FIX shutil.rmtree(dir_latest) BUG

+++ src/cbsl/edl/PageSearchResult.py
@@ -65,7 +65,8 @@ class PageSearchResult(WebpageWrapper):
                 log.debug(f'Wrote {file_name}')
 
         dir_latest = os.path.join(tempfile.gettempdir(), 'tmp.cbsl', 'latest')
-        shutil.rmtree(dir_latest)
+        if os.path.exists(dir_latest):
+            shutil.rmtree(dir_latest)
         shutil.copytree(dir_tmp_data, dir_latest)
         log.debug(f'Copied {dir_tmp_data} to {dir_latest}')
